requirements.txt
SpeechRecognition
setuptools
pyaudio
pyttsx3
openai
google-generativeai
requests
